requirements_win.txt
absl-py==2.1.0
accelerate==0.30.1
acme.hello==0.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
astor==0.8.1
asttokens==3.0.1
astunparse==1.6.3
asyncio==4.0.0
attrs==25.4.0
av==16.0.1
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.14.2
blake3==1.0.8
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
compressed-tensors==0.11.0
contourpy==1.3.3
crashtest==0.4.1
cryptography==46.0.3
cycler==0.12.1
dashscope==1.25.1
datasets==3.1.0
decorator==5.2.1
demjson3==3.0.6
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dm-env==1.6
dm-tree==0.1.9
dnspython==2.8.0
docstring_parser==0.17.0
dulwich==0.24.10
einops==0.8.1
email-validator==2.3.0
eval_type_backport==0.3.0
executing==2.2.1
fastapi==0.112.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.1
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2024.9.0
gast==0.6.0
gguf==0.17.1
google==3.0.0
google-auth==2.43.0
google-genai==1.52.0
google-pasta==0.2.0
grpcio==1.76.0
grpcio-tools==1.71.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
icecream==2.1.4
idna==3.11
importlib_metadata==8.7.0
installer==0.7.0
interegular==0.3.3
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
json5==0.10.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
keras==3.12.0
keyring==25.7.0
kiwisolver==1.4.9
langchain-core==1.1.1
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
lark==1.2.2
Levenshtein==0.27.1
libclang==18.1.1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.4
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral_common==1.8.5
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
namex==0.1.0
networkx==3.5
ninja==1.13.0
numba==0.61.2
numpy==1.26.4
openai==1.77.0
openai-harmony==0.0.8
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
ormsgpack==1.12.0
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post7
pbs-installer==2025.11.20
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.5.0
poetry==2.2.1
poetry-core==2.2.1
portpicker==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qwen-agent==0.0.16
qwen-vl-utils==0.0.11
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.45.0
setproctitle==1.3.7
shellingham==1.5.4
shtab==1.8.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.2.0
tiktoken==0.7.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.66.3
traitlets==5.14.3
transformers==4.50.0
transformers-stream-generator==0.0.4
trl==0.9.6
trove-classifiers==2025.11.14.15
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
ultralytics==8.3.129
ultralytics-thop==2.0.18
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.34.0
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
xgrammar==0.1.25
xxhash==3.6.0
yacs==0.1.8
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0

google-cloud-speech
pyaudio